move logic from Console to CommandParser
