+++ src/com/dd/command_util/CommandHandler.java
@@ -10,12 +10,8 @@ import com.dd.levels.Room;
 
 public abstract class CommandHandler {
 
-	public abstract void handleCommand(String[] args, CommandOutputLog outputLog) throws CommandHandlerException, FileNotFoundException;
-
-//	protected Player player = GameRunner.getActiveGameState().getActivePlayer();
-//	protected DungeonMap map = GameRunner.getActiveGameState().getMap();
-//	protected MapPosition playerPos = player.getPostion();
-//	protected Room currRoom = map.getRoom(playerPos);
+	//public abstract void handleCommand(String[] args, CommandOutputLog outputLog) throws CommandHandlerException, FileNotFoundException;
+	public abstract void handleCommand(String[] args) throws CommandHandlerException, FileNotFoundException;
 
     public class CommandHandlerException extends Exception {
         public CommandHandlerException(String message) {
