+++ src/com/dd/command_util/CommandParser.java
@@ -1,12 +1,15 @@
 package com.dd.command_util;
 
 import com.dd.command_util.CommandHandler;
+import com.dd.command_util.CommandHandler.CommandHandlerException;
 import com.dd.command_util.command.ExamineCommand;
 import com.dd.command_util.command.HelpCommand;
 import com.dd.command_util.command.MenuCommand;
 import com.dd.command_util.command.MoveCommand;
 import com.dd.command_util.command.QuitCommand;
 import com.dd.command_util.command.SaveCommand;
+
+import javafx.scene.control.TextInputControl;
 
 import java.io.FileNotFoundException;
 import java.lang.IllegalArgumentException;
@@ -16,6 +19,8 @@ import java.util.Map;
 public class CommandParser {
     private Map<String, CommandHandler> commandMap = new HashMap<String, CommandHandler>();
     private CommandOutputLog outputLog;
+    private String command = new String();
+	private String[] arguments = {"~!", null};
 
     public CommandParser(){
     	registerCommand("help", new HelpCommand());
@@ -25,6 +30,27 @@ public class CommandParser {
 		registerCommand("examine", new ExamineCommand());
 		registerCommand("save", new SaveCommand());
     }
+    
+    public void parse(String userInput) throws FileNotFoundException, CommandHandlerException {
+		CommandParser parser = new CommandParser();
+		String[] input = userInput.split(" ");
+		command = input[0].toLowerCase();
+		if(input.length > 1) {
+			arguments[0] = userInput.substring(command.length() + 1);
+			arguments[1] = null;
+		}
+		CommandHandler handler = commandMap.get(command);
+		if(handler == null)
+            throw new IllegalArgumentException("The command \""
+                                                    + command
+                                                    + "\" is invalid.");
+        try{
+            handler.handleCommand(arguments);
+        }
+        catch(CommandHandler.CommandHandlerException e){
+            throw e;
+        }
+	}
 
     public CommandParser(CommandOutputLog outputLog) {
         this.outputLog = outputLog;
@@ -51,18 +77,5 @@ public class CommandParser {
 
     public void setOutputLog(CommandOutputLog outputLog){
         this.outputLog = outputLog;
-    }
-
-    public void parseCommand(String command, String[] args) throws CommandHandler.CommandHandlerException, FileNotFoundException {
-        CommandHandler handler = commandMap.get(command);
-        if(handler == null)
-            throw new IllegalArgumentException("The command \""
-                                                    + command
-                                                    + "\" is invalid.");
-        try{
-            handler.handleCommand(args, outputLog);
-        }
-        catch(CommandHandler.CommandHandlerException e){
-            throw e;}
     }
 }
