+++ src/com/dd/command_util/command/HelpCommand.java
@@ -1,14 +1,14 @@
 package com.dd.command_util.command;
 
-import com.dd.Console;
 import com.dd.command_util.CommandHandler;
 import com.dd.command_util.CommandOutputLog;
+import com.dd.tester.Console;
 
 public class HelpCommand extends CommandHandler {
     public HelpCommand() {}
 	
 	@Override
-	public void handleCommand(String[] args, CommandOutputLog outputLog) throws CommandHandlerException {
+	public void handleCommand(String[] args) throws CommandHandlerException {
 		Console.updateScreen("\t\tAVAILABLE COMMANDS\n"
 				+ "\n"
 				+ "attack <entity_name>\n"
