+++ src/com/dd/command_util/command/UseCommand.java
@@ -7,7 +7,7 @@ public class UseCommand extends CommandHandler {
     public UseCommand() {}
 
     @Override
-    public void handleCommand(String[] args, CommandOutputLog outputLog){
+    public void handleCommand(String[] args){
 
     }
 }
